Added basic methods for fetching data from msD Moved Holder from Context in order to Spring cover more packages
