+++ src/test/cucumber/CheckMsdPax.feature
@@ -1,0 +1,10 @@
+Feature: Check msD booking passengers
+
+  Scenario: Check that passenger list is as expected
+    Given I should see that booking 96263099 has passengers list
+      | Humenskyi Vlady |
+      | Ivab Kuz        |
+      | Petro Klim      |
+      | Ivanka Klim     |
+      | Sofi Klim       |
+      | Alex Klim       |
