+++ src/test/java/com/thomascook/Context.java
@@ -1,0 +1,19 @@
+package com.thomascook;
+
+import org.springframework.context.annotation.*;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+
+@Configuration
+@PropertySources({
+        @PropertySource("classpath:config.properties"),
+        @PropertySource("classpath:credentials.properties")
+})
+@ComponentScan
+public class Context {
+
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer propertyConfigInDev() {
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+
+}
