+++ src/test/java/com/thomascook/ids/e2e/Holder.java
@@ -1,0 +1,68 @@
+package com.thomascook.ids.e2e;
+
+import com.thomascook.toscaAdapter.request.OTAPkgAvailRQ;
+import com.thomascook.toscaAdapter.response.OTAPkgAvailRS;
+import cucumber.api.DataTable;
+import org.opentravel.ota._2003._05.response.OTAPkgSearchRS;
+import org.springframework.stereotype.Component;
+
+import java.util.Properties;
+
+@Component
+public class Holder {
+
+    private OTAPkgSearchRS otaPkgSearchRS;
+    private String placeholder;
+    private static Properties properties;
+    private DataTable passengers;
+    private OTAPkgAvailRS toscaAvailabilityResponse;
+    private OTAPkgAvailRQ toscaAvailabiltyRequest;
+
+    public OTAPkgAvailRS getToscaAvailabilityResponse() {
+        return toscaAvailabilityResponse;
+    }
+
+    public void setToscaAvailabilityResponse(OTAPkgAvailRS toscaAvailabilityResponse) {
+        this.toscaAvailabilityResponse = toscaAvailabilityResponse;
+    }
+
+    public OTAPkgAvailRQ getToscaAvailabiltyRequest() {
+        return toscaAvailabiltyRequest;
+    }
+
+    public void setToscaAvailabiltyRequest(OTAPkgAvailRQ toscaAvailabiltyRequest) {
+        this.toscaAvailabiltyRequest = toscaAvailabiltyRequest;
+    }
+
+    public DataTable getPassengers() {
+        return passengers;
+    }
+
+    public void setPassengers(DataTable passengers) {
+        this.passengers = passengers;
+    }
+
+    public OTAPkgSearchRS getOtaPkgSearchRS() {
+        return otaPkgSearchRS;
+    }
+
+    public void setOtaPkgSearchRS(OTAPkgSearchRS otaPkgSearchRS) {
+        this.otaPkgSearchRS = otaPkgSearchRS;
+    }
+
+    public String getPlaceholder() {
+        return placeholder;
+    }
+
+    public void setPlaceholder(String placeholder) {
+        this.placeholder = placeholder;
+    }
+
+    public Properties getProperties() {
+        return properties;
+    }
+
+    public void setProperties(Properties properties) {
+        this.properties = properties;
+    }
+}
