+++ src/test/java/com/thomascook/ids/e2e/steps/Steps.java
@@ -1,12 +1,10 @@
 package com.thomascook.ids.e2e.steps;
 
-import com.thomascook.ids.e2e.Context;
 import com.thomascook.ids.e2e.tests.CreateBooking;
 import com.thomascook.nurvisAdapter.request.ReservationRequestTypeRequest;
 import com.thomascook.nurvisAdapter.response.ReservationResponseTypeResponse;
 import cucumber.api.java8.En;
 import org.opentravel.ota._2003._05.response.OTAPkgSearchRS;
-import org.springframework.test.context.ContextConfiguration;
 
 import javax.xml.bind.JAXBException;
 import javax.xml.datatype.DatatypeConfigurationException;
@@ -20,7 +18,7 @@ import static org.junit.Assert.assertNotNull;
 /**
  * Created by omm on 6/26/2017.
  */
-public class Steps implements En{
+public class Steps implements En {
 
     OTAPkgSearchRS otaPkgSearchRS;
     CreateBooking createBooking = new CreateBooking();
@@ -29,27 +27,28 @@ public class Steps implements En{
     static ReservationResponseTypeResponse nurvisBooking;
     static String onTourXML;
     static String destinationAirport;
+
     public Steps() {
 
         Given("^SOLR is requested for packages from (.*) airport to (.*) for (\\d+) adults$", (String fromAirport, String destination, Integer numberOfAdults) -> {
             // Write code here that turns the phrase above into concrete actions
             try {
                 otaPkgSearchRS = createBooking.getSOLRPackages(fromAirport, destination, numberOfAdults);
-            }  catch (JAXBException | IOException | DatatypeConfigurationException e) {
+            } catch (JAXBException | IOException | DatatypeConfigurationException e) {
                 e.printStackTrace();
             }
         });
 
         Given("^SOLR response has available packages$", () -> {
             // Write code here that turns the phrase above into concrete actions
-            assertNotEquals (otaPkgSearchRS.getHotelOffers().getStartIndex(),otaPkgSearchRS.getHotelOffers().getEndIndex());
+            assertNotEquals(otaPkgSearchRS.getHotelOffers().getStartIndex(), otaPkgSearchRS.getHotelOffers().getEndIndex());
         });
 
         When("^NURVIS is requested for booking$", () -> {
             // Write code here that turns the phrase above into concrete actions
             try {
                 //getValidPackage();
-                HashMap<String, Object> results =  getValidPackage(otaPkgSearchRS);
+                HashMap<String, Object> results = getValidPackage(otaPkgSearchRS);
                 assertNotNull(results);
                 nurvisRequest = (ReservationRequestTypeRequest) results.get("request");
                 nurvisInquiryResponse = (ReservationResponseTypeResponse) results.get("response");
@@ -65,8 +64,8 @@ public class Steps implements En{
         Then("^NURVIS returns booking ID$", () -> {
             // Write code here that turns the phrase above into concrete actions
             try {
-                nurvisBooking = createNurvisBooking(nurvisRequest,nurvisInquiryResponse);
-                assert(nurvisBooking.getFab().getWarning().get(0).getResultCode().equals("200"));
+                nurvisBooking = createNurvisBooking(nurvisRequest, nurvisInquiryResponse);
+                assert (nurvisBooking.getFab().getWarning().get(0).getResultCode().equals("200"));
                 destinationAirport = nurvisBooking.getFab().getDestination();
                 System.out.println(nurvisBooking.getFab().getWarning().get(0).getText().split(" ")[1]);
             } catch (JAXBException | IOException e) {
@@ -93,7 +92,7 @@ public class Steps implements En{
         When("^ontour xml is put on sftp$", () -> {
             // Write code here that turns the phrase above into concrete actions
             try {
-                assert(checkSFWForCustomer(nurvisBooking.getFab().getWarning().get(0).getText().split(" ")[1]));
+                assert (checkSFWForCustomer(nurvisBooking.getFab().getWarning().get(0).getText().split(" ")[1]));
             } catch (InterruptedException | IOException e) {
                 e.printStackTrace();
             }
@@ -102,7 +101,7 @@ public class Steps implements En{
 
         Then("^booking is processed$", () -> {
             // Write code here that turns the phrase above into concrete actions
-            assert(true);
+            assert (true);
         });
 
     }
